+++ tests/browser/incognito.test.js
@@ -1,24 +1,28 @@
 "use strict";
 
+const assert = require('assert');
 const Wendigo = require('../..');
 const configUrls = require('../config.json').urls;
 
+// This test does not explicitly check if the browser is in incognito mode as it cannot be realiably detected on a web
 describe("Incognito", function() {
-    this.timeout(50000);
+    this.timeout(5000);
 
-    it("Not Incognito", async() => {
-        const browser = await Wendigo.createBrowser();
-        await browser.open(configUrls.incognito);
-        await browser.wait(10);
-        await browser.assert.text("#check-text", "Not Incognito");
+    it("Incognito Browser", async() => {
+        const browser = await Wendigo.createBrowser({incognito: true});
+        await browser.open(configUrls.simple);
+        await browser.assert.text("p", "html_test");
+        assert.strictEqual(browser.settings.incognito, true);
+        assert.strictEqual(browser.incognito, true);
         await browser.close();
     });
 
-    it("Incognito", async() => {
-        const browser = await Wendigo.createBrowser({incognito: true});
-        await browser.open(configUrls.incognito);
-        await browser.wait(10);
-        await browser.assert.text("#check-text", "Incognito");
+    it("Not Incognito Browser", async() => {
+        const browser = await Wendigo.createBrowser();
+        await browser.open(configUrls.simple);
+        await browser.assert.text("p", "html_test");
+        assert.strictEqual(browser.settings.incognito, false);
+        assert.strictEqual(browser.incognito, false);
         await browser.close();
     });
 });